Mohamed Sprint 7 - First Push (#101)
* Database Schema Required Changes

Had to change the schema definition, not really a choice. The dummy data will probably be broken and need updating.

* Database Model Classes Updated

I updated the database model classes according to the revised schema. Labels are added to smplify the process of writing queries. Most of the SQL queries in the DatabaseQueryService class will need to be rewritten.

* Minor cleanup and renaming.

* All queries redone in DBQueryService.

- Proper MySql Connector Format used throughout. (single quotes around dates and string values, no semicolon at end of queries)
- Queries now reflect new database schema and new database model classes.